+++ console.cpp
@@ -1,20 +1,3 @@
-//	(c) Torbjörn Rathsman 2016
-//
-//	This file is part of Foobar.
-//
-//	Foobar is free software: you can redistribute it and/or modify
-//	it under the terms of the GNU General Public License as published by
-//	the Free Software Foundation, either version 3 of the License, or
-//	(at your option) any later version.
-//
-//	Foobar is distributed in the hope that it will be useful,
-//	but WITHOUT ANY WARRANTY; without even the implied warranty of
-//	MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
-//	GNU General Public License for more details.
-//
-//	You should have received a copy of the GNU General Public License
-//	along with Foobar.  If not, see <http://www.gnu.org/licenses/>.
-
 #ifdef __WAND__
 target[name[console.o] type[object]]
 #endif
@@ -253,7 +236,7 @@ Console& Console::write(uint32_t codepoint)
 		{
 		case 65535:
 			m_position=std::max(m_position
-				,m_n_cols*(m_position/m_n_cols) + 1lu);
+				,m_n_cols*(m_position/m_n_cols) + static_cast<size_t>(1));
 			while((m_position%m_n_cols)!=0)
 				{
 				write(' ');
@@ -276,7 +259,7 @@ Console& Console::write(uint32_t codepoint)
 			m_colors_bg[4*position + 2]=color_bg;
 			m_colors_bg[4*position + 3]=color_bg;
 
-			position=(position+1)%N;
+			position=(position + 1)%N;
 
 			if(position%colsCountGet()==0)
 				{++m_line_current;}
